+++ gnerd/Front/gcobros_web/src/helper/jwt.js
@@ -28,10 +28,6 @@ export function getAdminData() {
   }
 }
 
-export function jwtLogout() {
-  localStorage.removeItem('token');
-}
-
 export const jwt = () => localStorage.getItem('token') || '';
 
 export function verifyJWT(token) {
